refactor move baseFormStateTypes to store/editor/ directory
